MCP/Ecommerce_MCP_Server/ecommerce-mcp-cdk/requirements.txt
# AWS CDK dependencies
aws-cdk-lib>=2.100.0
constructs>=10.0.0

# AWS SDK for custom resources and data loading
boto3>=1.34.0
botocore>=1.34.0

# Synthetic data generation
faker>=24.0.0

# Utilities
python-dateutil>=2.8.2
